requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy
opencv-python
tqdm
pyyaml
onnx
onnxruntime-gpu